refactor: move GameCardFragment to presentation layer
