+++ app/src/main/res/layout/fragment_game_card.xml
@@ -12,7 +12,7 @@
         android:layout_width="match_parent"
         android:layout_height="match_parent"
         android:background="@color/surface"
-        tools:context=".view.GameCardFragment">
+        tools:context=".presentation.view.GameCardFragment">
 
         <androidx.recyclerview.widget.RecyclerView
             android:id="@+id/recyclerView"
